+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to the "printcode" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## 1.1.0
+## 2.0.0
 
-- Added support for browserPath in settings. 
+- Added support for browserPath in settings.
+- Fixed for external CDN dependency.
+- Added Google AdSense.
+- Added Google Analytics.
 
 ## 1.0.0
 - Initial release